+++ packages/azerothjs/src/index.ts
@@ -44,6 +44,7 @@ export {
     untrack,
     on,
     onCleanup,
+    onMount,
     onRootDispose,
     createRoot,
     getOwner,